% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainmethod.R
\name{generateData}
\alias{generateData}
\title{Generate example data}
\usage{
generateData()
}
\arguments{
\item{None}{}
}
\value{
output A list of four datasets
}
\description{
This method generates data for 4 hospitals and save each dataset
in its own directory locally.
}
\examples{
generateData()

}
\keyword{data}
\keyword{simulate}
