% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainmethod.R
\name{getLocalodds.plot}
\alias{getLocalodds.plot}
\title{Plot local log odds ratios and their variances on a forest plot.}
\usage{
getLocalodds.plot()
}
\value{
A forest plot that shows local odds ratios and their variances and meta-analytical results
}
\description{
Plot local log odds ratios and their variances on a forest plot.
}
\examples{
getLocalodds.plot()

}
