% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainmethod.R
\name{getLocalodds}
\alias{getLocalodds}
\title{Using local data to obtain log odds ratio and variance estimates for each hospital}
\usage{
getLocalodds()
}
\arguments{
\item{None}{}
}
\value{
A list that includes local odds ratio for each hospital and their variances.
Meta log odds ratio and its variance.
}
\description{
This function calculates a local log odds ratio and their variance estimates.
With the local log odds ratio and their variance estimates, the function also
compute a variance-inverted meta-analytical log odds ratio and variance estimates
by any spaces.
}
\examples{
getLocalodds()
}
\keyword{Local}
\keyword{calculation}
