% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainmethod.R
\name{Update_ate}
\alias{Update_ate}
\title{Update causal log odds ratio estimates}
\usage{
Update_ate(
  beta,
  A_all,
  y,
  X,
  y_treated_sum,
  y_control_sum,
  n_treated,
  n_control,
  lastsite = FALSE
)
}
\arguments{
\item{beta}{causal log odds ratio from the previous site}

\item{A_all}{local treatment assignment}

\item{y}{local outcome data.}

\item{X}{local covariates data.}

\item{n_treated}{}

\item{n_control}{}

\item{lastsite}{defult is false which indicates whether the current site is the \code{lastsite}}
}
\value{
a vector of intermediate quantites that will be passed down to the next site when \code{lastsite} is FALSE;
a vector of causal log odds ratio and its variance when \code{lastsite} is TRUE
}
\description{
This function updates estimates for the collaborative causal log odds ratio and its variance.
}
\examples{
Update_ate(beta,A_all,y,X,y_treated_sum,y_control_sum,n_treated,n_control,TRUE)
}
\keyword{ATE}
\keyword{Second}
\keyword{round}
\keyword{update}
