% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainmethod.R
\name{Update_beta}
\alias{Update_beta}
\title{Update propensity score estimates}
\usage{
Update_beta(X, y, betahat, sum2, type = "binomial")
}
\arguments{
\item{X}{local covariates data.}

\item{y}{local outcome data.}

\item{betahat}{current estimates for propensity scores which is usually passed down from the preceding site.}

\item{sum2}{current Hessian matrix which is usually passed down from the preceding site.}
}
\value{
the updated \code{betahat} and \code{sum2}.
}
\description{
This function updates estimates for propensity scores and their vairances.
}
\examples{
Update_beta(X,y, betahat, sum2,type)
}
\keyword{First}
\keyword{round}
\keyword{update}
